pyproject.toml
[project]
name = "learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent>=0.1.3",
    "openai>=1.95.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
]

[tool.uv.workspace]
members = [
    "myproject",
]
